Feature/routes error handling

+++ services/incidentService.js
@@ -2,28 +2,6 @@ const { Incident, User, Item } = require("../models/");
 const sequelize = require("sequelize");
 const Op = sequelize.Op;
 
-//SIN TERMINAR!
-
-exports.getIncident = async (id) => {
-  const incident = await Incident.findByPk(id, {
-    include: [
-      {
-        association: Incident.Item,
-      },
-    ],
-  });
-  return incident;
-};
-
-exports.createIncident = async (incident) => {
-  const newIncident = await Incident.create(incident);
-  return newIncident;
-};
-
-exports.editIncident = (id, body) => {
-  return Incident.findByPk(id).then((incident) => incident.update(body));
-};
-
 exports.getAllIncidents = async () => {
   const incidents = await Incident.findAll({
     include: [
@@ -32,20 +10,42 @@ exports.getAllIncidents = async () => {
       },
     ],
   });
+  if (!incidents) throw 404;
   return incidents;
+};
+
+exports.createIncident = async (incident) => {
+  if (Object.keys(incident).length === 0) return 400;
+  const newIncident = await Incident.create(incident);
+  return newIncident;
+};
+
+exports.editIncident = async (id, body) => {
+  if (isNaN(id)) throw 400;
+  const incident = await Incident.findByPk(id);
+  if (!incident) throw 404;
+  if (Object.keys(body).length === 0) return 400;
+  await incident.update(body);
+  return incident;
 };
 
 exports.getByUserId = async (userId) => {
+  if (isNaN(userId)) throw 400;
   const incidents = await Incident.findAll({ where: { userId: userId } });
+  if (!incidents.length) throw 404;
   return incidents;
 };
 
-exports.deleteIncident = (id) => {
+exports.deleteIncident = async (id) => {
+  if (isNaN(id)) throw 400;
+  const incident = await Incident.findByPk(id);
+  if (!incident) throw 404;
   return Incident.destroy({ where: { id } });
 };
 
 exports.getSearchedIncidents = async (filter) => {
-  if (!isNumeric(filter)) {
+  if (!filter) throw 400;
+  if (isNaN(filter)) {
     const results = await Incident.findAll({
       where: {
         status: filter,
@@ -56,6 +56,7 @@ exports.getSearchedIncidents = async (filter) => {
         },
       ],
     });
+    if (!results.length) throw 404;
     return results;
   } else {
     const results = await Incident.findAll({
@@ -68,11 +69,7 @@ exports.getSearchedIncidents = async (filter) => {
         },
       ],
     });
+    if (!results.length) throw 404;
     return results;
   }
 };
-
-function isNumeric(str) {
-  if (typeof str != "string") return false;
-  return !isNaN(str) && !isNaN(parseFloat(str));
-}
